EN v3.4.1 EN profile scaling integration with isHigh logic bugfix

+++ core/interfaces/src/main/kotlin/app/aaps/core/interfaces/aps/ENConfig.kt
@@ -25,7 +25,7 @@ data class ENConfig(
     var ENWEndTime: Long?,
     var ENWRunTime: Int,
     var ENWDuration: Int,
-    var enwProfileScalePct: Int,
+    var enwProfileScalePct: Double,
     var ENWcobMaxbolus: Double,
     var ENWuamMaxbolus: Double,
     var ENWNetIOB: Double,
